# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tag {
  id: number;
  name: string;
  color: string;
  createdAt: string;
}

export interface CreateTagBody {
  name: string;
  color: string;
}

export interface UpdateTagBody {
  name?: string;
  color?: string;
}

export interface SetPlaceTagsBody {
  tagIds: number[];
}

export interface SavedPlace {
  id: number;
  name: string;
  address: string;
  latitude: number;
  longitude: number;
  category?: string | null;
  note?: string | null;
  /** Estimated visit duration in minutes */
  estimatedDuration?: number | null;
  /** Estimated cost in USD */
  estimatedCost?: number | null;
  /** Priority 1-5 */
  priority?: number | null;
  openingHours?: string | null;
  createdAt: string;
  tags: Tag[];
}

export interface CreateSavedPlaceBody {
  name: string;
  address: string;
  latitude: number;
  longitude: number;
  category?: string | null;
  note?: string | null;
  estimatedDuration?: number | null;
  estimatedCost?: number | null;
  priority?: number | null;
  openingHours?: string | null;
}

export interface UpdateSavedPlaceBody {
  name?: string;
  address?: string;
  latitude?: number;
  longitude?: number;
  category?: string | null;
  note?: string | null;
  estimatedDuration?: number | null;
  estimatedCost?: number | null;
  priority?: number | null;
  openingHours?: string | null;
}

export interface Trip {
  id: number;
  name: string;
  startDate: string;
  endDate: string;
  description?: string | null;
  totalDays: number;
  createdAt: string;
}

export interface TransportMode {
  id: number;
  name: string;
  icon: string;
  color: string;
  createdAt: string;
}

export interface ItineraryItem {
  id: number;
  tripId: number;
  savedPlaceId?: number | null;
  dayNumber: number;
  position: number;
  customName?: string | null;
  note?: string | null;
  estimatedDuration?: number | null;
  estimatedCost?: number | null;
  startMinutes?: number | null;
  transportModeId?: number | null;
  transportMode?: { id: number; name: string; icon: string; color: string } | null;
  place?: SavedPlace | null;
}

export interface TripDay {
  dayNumber: number;
  date: string;
  items: ItineraryItem[];
}

export interface TripDetail {
  id: number;
  name: string;
  startDate: string;
  endDate: string;
  description?: string | null;
  totalDays: number;
  createdAt: string;
  days: TripDay[];
}

export interface CreateTripBody {
  name: string;
  startDate: string;
  endDate: string;
  description?: string | null;
}

export interface UpdateTripBody {
  name?: string;
  startDate?: string;
  endDate?: string;
  description?: string | null;
}

export interface CreateItineraryItemBody {
  savedPlaceId: number;
  note?: string | null;
  estimatedDuration?: number | null;
  estimatedCost?: number | null;
  startMinutes?: number | null;
  transportModeId?: number | null;
}

export interface UpdateItineraryItemBody {
  note?: string | null;
  estimatedDuration?: number | null;
  estimatedCost?: number | null;
  dayNumber?: number | null;
  startMinutes?: number | null;
  transportModeId?: number | null;
}

export interface CreateTransportModeBody {
  name: string;
  icon: string;
  color: string;
}

export interface UpdateTransportModeBody {
  name?: string;
  icon?: string;
  color?: string;
}

export interface ReorderItineraryItemsBody {
  /** Ordered array of item IDs representing the new order */
  itemIds: number[];
}

export type BudgetItemCategory =
  (typeof BudgetItemCategory)[keyof typeof BudgetItemCategory];

export const BudgetItemCategory = {
  transportation: "transportation",
  accommodation: "accommodation",
  food: "food",
  attractions: "attractions",
  shopping: "shopping",
  miscellaneous: "miscellaneous",
} as const;

export interface BudgetItem {
  id: number;
  tripId: number;
  category: BudgetItemCategory;
  description: string;
  amount: number;
  dayNumber?: number | null;
  createdAt: string;
}

export type CreateBudgetItemBodyCategory =
  (typeof CreateBudgetItemBodyCategory)[keyof typeof CreateBudgetItemBodyCategory];

export const CreateBudgetItemBodyCategory = {
  transportation: "transportation",
  accommodation: "accommodation",
  food: "food",
  attractions: "attractions",
  shopping: "shopping",
  miscellaneous: "miscellaneous",
} as const;

export interface CreateBudgetItemBody {
  category: CreateBudgetItemBodyCategory;
  description: string;
  amount: number;
  dayNumber?: number | null;
}

export type UpdateBudgetItemBodyCategory =
  (typeof UpdateBudgetItemBodyCategory)[keyof typeof UpdateBudgetItemBodyCategory];

export const UpdateBudgetItemBodyCategory = {
  transportation: "transportation",
  accommodation: "accommodation",
  food: "food",
  attractions: "attractions",
  shopping: "shopping",
  miscellaneous: "miscellaneous",
} as const;

export interface UpdateBudgetItemBody {
  category?: UpdateBudgetItemBodyCategory;
  description?: string;
  amount?: number;
  dayNumber?: number | null;
}

export type BudgetSummaryCategoryTotals = { [key: string]: number };

export type BudgetSummaryDailyTotalsItem = {
  dayNumber: number;
  total: number;
};

export interface BudgetSummary {
  tripId: number;
  totalBudget: number;
  categoryTotals: BudgetSummaryCategoryTotals;
  dailyTotals: BudgetSummaryDailyTotalsItem[];
  /** Sum of estimated costs from itinerary items */
  placesCost: number;
}