# lib/api-zod/src/generated/types/budgetItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetItemCategory } from "./budgetItemCategory";

export interface BudgetItem {
  id: number;
  tripId: number;
  category: BudgetItemCategory;
  description: string;
  amount: number;
  dayNumber?: number | null;
  createdAt: Date;
}