# lib/api-zod/src/generated/types/createItineraryItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateItineraryItemBody {
  savedPlaceId: number;
  note?: string | null;
  estimatedDuration?: number | null;
  estimatedCost?: number | null;
}