# lib/api-zod/src/generated/types/itineraryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { SavedPlace } from "./savedPlace";

export interface ItineraryItem {
  id: number;
  tripId: number;
  savedPlaceId?: number | null;
  dayNumber: number;
  position: number;
  customName?: string | null;
  note?: string | null;
  estimatedDuration?: number | null;
  estimatedCost?: number | null;
  place?: SavedPlace | null;
}