# lib/api-zod/src/generated/types/updateBudgetItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBudgetItemBodyCategory } from "./updateBudgetItemBodyCategory";

export interface UpdateBudgetItemBody {
  category?: UpdateBudgetItemBodyCategory;
  description?: string;
  amount?: number;
  dayNumber?: number | null;
}